Disable a check if multimedia msg have more than 1 media
According to the tg docs msg sent with send_media_group has to have more than
1 media attached. However it seems like tg/telebot does handle correctly a
media group msg with just 1 media attached so I decided to disable that check.

As for now I'm gonna leave the whole logic for original msgs with just 1 media
in place and the only possible scenario when media gorup msg can have 1 media
is when some api returns more than 1 media, but all but 1 media are in
an unsupported format or when the post has more than 10 media and after split
the last msg ends up with just 1 media. The check in question applied to the
first scenario (some media were in an unsupported format) so as it is a very
rare scenario I decided it just doesn't matter and as long as it works I will
leave it as is.